+++ frontend/lost/src/containers/Annotation/SIA/lost-sia/src/Canvas.js
@@ -54,6 +54,12 @@ import InfoBoxes from './InfoBoxes/InfoBoxArea'
  *              polygons: []
  *          }
  *      }
+ * @param {object} annoSaveResponse - Backend response when updating an annotation in backend
+ *                  {
+ *                      tempId: int or str, // temporal frontend Id 
+ *                      dbId: int, // Id from backend
+ *                      newStatus: str // new Status for the annotation
+ *                  }
  * @param {object} possibleLabels - Possible labels that can be assigned to 
  *      an annotation.
  *      {   
@@ -101,17 +107,21 @@ import InfoBoxes from './InfoBoxes/InfoBoxArea'
  *          }
  *      },
  *      allowedToMarkExample: bool, -> Indicates wether the current user is allowed to mark an annotation as example.
- *      autoSaveInterval: int -> Interval in seconds when an autosave will be requested by canvas
  *   }
  * @param {str or int} defaultLabel (optional) - Name or ID of the default label that is used
  *      when no label was selected by the annotator. If not set "no label" will be used.
  *      If ID is used, it needs to be one of the possible label ids.
  * @param {bool} blocked Block canvas view with loading dimmer.
  * @param {bool} preventScrolling Prevent scrolling on mouseEnter
- * @param {int} nextAnnoId Id that will be used for the next annotation that 
- *        will be created. If undefined, the canvas will create its own ids.
  * @param {bool} lockedAnnos A list of AnnoIds of annos that should only be displayed.
  *      Such annos can not be edited in any way.
+ * @event onAnnoSaveEvent - Callback with update information for a single 
+ *          annotation or the current image that can be used for backend updates
+ *          args: {
+ *                      action: the action that was performed in frontend, 
+ *                      anno: anno information, 
+ *                      img: image information
+ *              }
  * @event onNotification - Callback for Notification messages
  *      args: {title: str, message: str, type: str}
  * @event onKeyDown - Fires for keyDown on canvas 
@@ -157,7 +167,6 @@ class Canvas extends Component{
             },
             annos: [],
             mode: modes.VIEW,
-            // selectedAnnoId: {id:undefined},
             selectedAnnoId: undefined,
             showSingleAnno: undefined,
             showLabelInput: false,
@@ -175,7 +184,6 @@ class Canvas extends Component{
             annoToolBarVisible: false,
             possibleLabels: undefined,
             annoCommentInputTrigger: 0,
-            // showAnnoCommentInput: false
         }
         this.img = React.createRef()
         this.svg = React.createRef()
@@ -184,7 +192,6 @@ class Canvas extends Component{
         this.keyMapper = new KeyMapper((keyAction) => this.handleKeyAction(keyAction))
         this.mousePosAbs = undefined
         this.clipboard = undefined
-        this.autoSaveI = undefined
     }
 
     componentDidMount(){
@@ -193,13 +200,6 @@ class Canvas extends Component{
 
             this.setState({prevLabel:[this.props.defaultLabel]})
         }
-        if (this.props.canvasConfig.autoSaveInterval){
-            this.autoSaveI = setInterval(() => {
-                // console.log('AutoSave')
-                // this.props.onAutoSave()
-                this.triggerCanvasEvent(canvasActions.CANVAS_AUTO_SAVE)
-            }, this.props.canvasConfig.autoSaveInterval*1000)
-        }
         if (this.props.onGetFunction){
             this.props.onGetFunction({
                 'deleteAllAnnos':() => this.deleteAllAnnos(),
@@ -207,28 +207,19 @@ class Canvas extends Component{
                 'resetZoom': () => this.resetZoom(),
                 'getAnnos': (annos,removeFrontendIds) => this.getAnnos(annos, removeFrontendIds)
             })
-            // this.props.onGetFunction(
-            //     () => this.deleteAllAnnos()
-            // )
         }
     }
 
-    componentWillUnmount(){
-        clearInterval(this.autoSaveI)
-        console.log('Cleard auto save interval!')
-    }
-
     componentDidUpdate(prevProps, prevState){
-        // if (this.props.image.id !== prevProps.image.id){
-            
-        // }
+        if (prevProps.annoSaveResponse !== this.props.annoSaveResponse){
+            this.updateAnnoBySaveResponse(this.props.annoSaveResponse)
+        }
         if (prevProps.imageMeta !== this.props.imageMeta){
             if (this.props.imageMeta){
                 this.setState({
                     imgLabelIds: this.props.imageMeta.labelIds,
                     imgAnnoTime: this.props.imageMeta.annoTime,
                     imgLoadTimestamp: performance.now()
-                    // isJunk: this.props.annos.image.isJunk
                 })
             }
         }
@@ -238,31 +229,21 @@ class Canvas extends Component{
                     annoConversion.fixBackendAnnos(this.props.annos) 
                 ) 
             }
-            // this.setState({
-            //     imageLoaded: false,
-            //     // imageBlob: undefined
-            // })
         }
         if (prevProps.isJunk !== this.props.isJunk){
             if (this.state.isJunk !== this.props.isJunk){
                 this.setState({
                     isJunk: this.props.isJunk
                 })
+                if (this.state.imageLoaded){
+                    this.handleAnnoSaveEvent(canvasActions.IMG_JUNK_UPDATE, undefined, 
+                        {isJunk:this.props.isJunk})
+                }
             }
         }
         if (this.state.imageBlob !== this.props.imageBlob){
             this.setState({imageBlob: this.props.imageBlob})
         }
-        // if (!this.state.imageLoaded){
-        //     if(this.props.annos.image.id === this.props.image.id){
-        //         this.setState({
-        //             imageLoaded: true
-        //         })
-        //         if (this.props.onImageLoaded){
-        //             this.props.onImageLoaded()
-        //         }
-        //     }
-        // }
         if (this.props.possibleLabels !== prevProps.possibleLabels){
             this.updatePossibleLabels()
         }
@@ -297,8 +278,6 @@ class Canvas extends Component{
             } 
             if(prevProps.layoutUpdate !== this.props.layoutUpdate){
                 this.selectAnnotation(undefined)
-                // this.updateCanvasView(this.getAnnoBackendFormat())
-                // const {imageOffset} = this.updateImageSize()
                 this.updateCanvasView(annoConversion.canvasToBackendAnnos(
                     this.state.annos, this.state.svg, false, this.state.imageOffset
                 ))
@@ -316,9 +295,6 @@ class Canvas extends Component{
             selectedAnnoId: undefined
         })
         this.triggerCanvasEvent(canvasActions.CANVAS_IMG_LOADED)
-        // if (this.props.onImageLoaded){
-        //     this.props.onImageLoaded()
-        // }
     }
 
     onMouseOver(){
@@ -416,20 +392,10 @@ class Canvas extends Component{
         }
     }
 
-    // handleAnnoCommentInputClose(comment){
-    //     const anno = this.findAnno(this.state.selectedAnnoId)
-    //     anno.comment = comment
-    //     this.stopAnnotimeMeasure(anno)
-    //     this.updateSelectedAnno(anno, modes.VIEW)
-    //     console.log('handleAnnoCommentInputClose', comment)
-    // }
-    
     updateAnnoComment(comment){
         const anno = this.findAnno(this.state.selectedAnnoId)
         anno.comment = comment
         this.handleAnnoEvent(anno, canvasActions.ANNO_COMMENT_UPDATE)
-        // this.stopAnnotimeMeasure(anno)
-        // console.log('updateAnnoComment', comment)
     }
 
     handleKeyAction(action){
@@ -455,8 +421,6 @@ class Canvas extends Component{
             case keyActions.TOGGLE_ANNO_COMMENT_INPUT:
                 if (this.state.selectedAnnoId) {
                     this.setState({annoCommentInputTrigger: this.state.annoCommentInputTrigger+1})
-                    // this.startAnnotimeMeasure(anno)
-                    // this.showSingleAnno(this.state.selectedAnnoId)
                 }
                 break
             case keyActions.DELETE_ANNO_IN_CREATION:
@@ -467,7 +431,6 @@ class Canvas extends Component{
                     this.updateSelectedAnno(
                         anno, modes.ADD
                     )
-                    // this.showSingleAnno(anno.id)
                 }
                 break
             case keyActions.LEAVE_ANNO_ADD_MODE:
@@ -475,7 +438,6 @@ class Canvas extends Component{
                     this.updateSelectedAnno(
                         anno, modes.VIEW
                     )
-                    // this.showSingleAnno(undefined)
                 }
                 break
             case keyActions.UNDO:
@@ -488,7 +450,6 @@ class Canvas extends Component{
                 this.traverseAnnos()
                 break
             case keyActions.CAM_MOVE_LEFT:
-                // this.setMode(modes.CAMERA_MOVE)
                 this.moveCamera(camKeyStepSize, 0)
                 break
             case keyActions.CAM_MOVE_RIGHT:
@@ -501,7 +462,6 @@ class Canvas extends Component{
                 this.moveCamera(0, -camKeyStepSize)
                 break
             case keyActions.CAM_MOVE_STOP:
-                // this.setMode(modes.VIEW)
                 break
             case keyActions.COPY_ANNOTATION:
                 this.copyAnnotation()
@@ -522,7 +482,6 @@ class Canvas extends Component{
     onKeyDown(e){
         e.preventDefault()
         this.keyMapper.keyDown(e.key)
-        // this.findAnno(this.state.selectedAnnoId)
         if (this.props.onKeyDown){
             this.props.onKeyDown(e)
         }
@@ -566,6 +525,8 @@ class Canvas extends Component{
     handleAnnoEvent(anno, pAction){
         let newAnnos = undefined
         switch(pAction){
+            case canvasActions.ANNO_ENTER_CREATE_MODE:
+                break
             case canvasActions.ANNO_MARK_EXAMPLE:
                 newAnnos = this.updateSelectedAnno(anno, modes.VIEW)
                 this.pushHist(
@@ -596,6 +557,7 @@ class Canvas extends Component{
                 )
                 this.showSingleAnno(undefined)
                 this.setState({annoToolBarVisible:true})
+                this.handleAnnoSaveEvent(pAction, anno, undefined)
                 break
             case canvasActions.ANNO_MOVED:
                 anno = this.stopAnnotimeMeasure(anno)
@@ -606,6 +568,7 @@ class Canvas extends Component{
                     pAction, undefined
                 )
                 this.setState({annoToolBarVisible:true})
+                this.handleAnnoSaveEvent(pAction, anno, undefined)
                 break
             case canvasActions.ANNO_ENTER_MOVE_MODE:
                 anno = this.startAnnotimeMeasure(anno)
@@ -625,6 +588,7 @@ class Canvas extends Component{
                     newAnnos, anno.id,
                     pAction, this.state.showSingleAnno
                 )
+                this.handleAnnoSaveEvent(pAction, anno, undefined)
                 break
             case canvasActions.ANNO_REMOVED_NODE:
                 if (!this.checkAnnoLength(anno)){
@@ -637,6 +601,7 @@ class Canvas extends Component{
                     newAnnos, anno.id,
                     pAction, this.state.showSingleAnno
                 )
+                this.handleAnnoSaveEvent(pAction, anno, undefined)
                 break
             case canvasActions.ANNO_EDITED:
                 anno = this.stopAnnotimeMeasure(anno)
@@ -646,21 +611,22 @@ class Canvas extends Component{
                     pAction, this.state.showSingleAnno
                 )
                 this.setState({annoToolBarVisible:true})
+                this.handleAnnoSaveEvent(pAction, anno, undefined)
                 break
             case canvasActions.ANNO_DELETED:
-                newAnnos = this.updateSelectedAnno(anno, modes.DELETED)
+                const res = this.updateSelectedAnno(anno, modes.DELETED, true)
+                newAnnos = res.newAnnos
                 this.selectAnnotation(undefined)
                 this.showSingleAnno(undefined)
                 this.pushHist(
                     newAnnos, undefined,
                     pAction, this.state.showSingleAnno
                 )
+                this.handleAnnoSaveEvent(pAction, res.newAnno, undefined)
                 break
             case canvasActions.ANNO_COMMENT_UPDATE:
-                newAnnos = this.updateSelectedAnno(anno, modes.VIEW)
-                // newAnnos = this.updateSelectedAnno(anno, modes.DELETED) 
-                // this.selectAnnotation(undefined)
-                // this.showSingleAnno(undefined)
+                const res_comment = this.updateSelectedAnno(anno, modes.VIEW, true)
+                newAnnos = res_comment.newAnnos
                 this.pushHist(
                     newAnnos, anno.id,
                     pAction, this.state.showSingleAnno
@@ -670,6 +636,7 @@ class Canvas extends Component{
                     message: `Saved comment: ${anno.comment}`,
                     type: notificationType.SUCCESS
                 })
+                this.handleAnnoSaveEvent(pAction, res_comment.newAnno, undefined)
                 break
             case canvasActions.ANNO_LABEL_UPDATE:
                 anno = this.stopAnnotimeMeasure(anno)
@@ -683,6 +650,7 @@ class Canvas extends Component{
                     pAction, undefined
                 )
                 this.setState({annoToolBarVisible:true})
+                this.handleAnnoSaveEvent(pAction, anno, undefined)
                 break
             case canvasActions.ANNO_CREATED_NODE:
                 anno = this.stopAnnotimeMeasure(anno)
@@ -701,13 +669,30 @@ class Canvas extends Component{
                 )
                 this.showSingleAnno(undefined)
                 this.setState({annoToolBarVisible:true})
+                this.handleAnnoSaveEvent(pAction, anno, undefined)
                 break
             default:
-                console.warn('Action not handeled', pAction)
+                console.warn('Action not handled', pAction)
                 break
         }
         if (this.props.onAnnoEvent){
             this.props.onAnnoEvent(anno, newAnnos, pAction)
+        }
+    }
+
+    handleAnnoSaveEvent(action, anno, img){
+        const imgData = {...img,
+            imgId: this.props.imageMeta.id,
+            annoTime: this.props.imageMeta.annoTime + (performance.now() - this.state.imgLoadTimestamp)/1000
+        }
+        const saveData = {
+            anno: anno ? annoConversion.canvasToBackendSingleAnno(anno, this.state.svg, 
+                false, this.state.imageOffset): undefined,
+            img: imgData,
+            action
+            }
+        if (this.props.onAnnoSaveEvent){
+            this.props.onAnnoSaveEvent(saveData)
         }
     }
 
@@ -725,22 +710,34 @@ class Canvas extends Component{
 
     handleImgBarClose(){
         this.triggerCanvasEvent(canvasActions.CANVAS_IMGBAR_CLOSE)
-        // if (this.props.onImgBarClose){
-        //     this.props.onImgBarClose()
-        // }
+    }
+
+    gotNewLabel(label){
+        let ret = false
+        label.forEach(e => {
+            if (!this.state.imgLabelIds.includes(e)) ret = true
+        })
+        return ret
     }
 
     handleImgLabelUpdate(label){
-        this.setState({
-            imgLabelIds: label,
-            imgLabelChanged: true,
-        })
-        this.pushHist(this.state.annos,
-            this.state.selectedAnnoId,
-            canvasActions.IMG_LABEL_UPDATE,
-            this.state.showSingleAnno,
-            label
-        )
+        if (this.gotNewLabel(label)){
+            this.setState({
+                imgLabelIds: label,
+                imgLabelChanged: true,
+            })
+            this.pushHist(this.state.annos,
+                this.state.selectedAnnoId,
+                canvasActions.IMG_LABEL_UPDATE,
+                this.state.showSingleAnno,
+                label
+            )
+            const imgData = {
+                imgLabelIds: label,
+                imgLabelChanged: true
+            }
+            this.handleAnnoSaveEvent(canvasActions.IMG_LABEL_UPDATE, undefined, imgData)
+        }
     }
 
     handleCanvasClick(e){
@@ -755,14 +752,7 @@ class Canvas extends Component{
 
     handleImgLabelInputClose(){
         this.triggerCanvasEvent(canvasActions.CANVAS_LABEL_INPUT_CLOSE)
-        // if (this.props.onImgLabelInputClose){
-        //     this.props.onImgLabelInputClose()
-        // }
     }
-
-    // handleAnnoCommentInputClose(e){
-    //     this.showAnnoCommentInputField(false)
-    // }
 
     handleSvgMouseMove(e){
         this.mousePosAbs = mouse.getMousePositionAbs(e, this.state.svg)
@@ -775,10 +765,8 @@ class Canvas extends Component{
     }
 
     handleHideLbl(lbl, hide){
-        // console.log('hide lbl', lbl, hide)
         let hiddenSelected = false
         const newAnnos = this.state.annos.map(anno => {
-            // console.log('map annos', anno)
             const newAnno = {...anno}
             if (anno.labelIds.includes(lbl.id)){
                 newAnno.visible = !hide
@@ -791,7 +779,6 @@ class Canvas extends Component{
             }
             return newAnno
         })
-        // console.log('newAnnos', newAnnos)
         this.setState({annos: newAnnos})
         if (hiddenSelected){
             this.selectAnnotation(undefined)
@@ -855,14 +842,6 @@ class Canvas extends Component{
             })
             return false
         } 
-        // if (anno.type === 'line' && anno.data.length < 2){
-        //     this.handleNotification({
-        //         title: "Invalid line!",
-        //         message: 'A vaild line needs at least 2 points!',
-        //         type: notificationType.WARNING
-        //     })
-        //     return false
-        // } 
         return true
     }
 
@@ -914,9 +893,11 @@ class Canvas extends Component{
         }
     }
     unloadImage(){
+        console.log('unloadImage', this.state, this.props.imageMeta)
         if(this.state.imageLoaded){
             this.setState({imageLoaded:false})
         }
+        this.handleAnnoSaveEvent(canvasActions.IMG_ANNO_TIME_UPDATE, undefined, undefined)
     }
     /**
      * Find a annotation by id in current state
@@ -986,10 +967,7 @@ class Canvas extends Component{
     deleteAnnoInCreationMode(anno) {
         if (anno){
             if (anno.mode === modes.CREATE){
-                // const ar = this.findAnnoRef(this.state.selectedAnnoId)
-                // if (ar !== undefined) ar.current.myAnno.current.removeLastNode()
                 this.handleAnnoEvent(anno, canvasActions.ANNO_DELETED)
-    
             } else {
             }
         }
@@ -999,13 +977,10 @@ class Canvas extends Component{
         let newAnnos = []
         this.state.annos.forEach( e => {
             if ((typeof e.id) !== "string"){
-                newAnnos.push(
-                    {...e, status: annoStatus.DELETED}
-                )
+                const anno = {...e, status: annoStatus.DELETED}
+                this.handleAnnoEvent(anno, canvasActions.ANNO_DELETED)
             }
         })
-        this.pushHist(newAnnos, undefined, 'deletedAllAnnos', this.state.showSingleAnno, this.state.imgLabelIds)
-        this.setState({annos: newAnnos})
         this.selectAnnotation(undefined)
         this.showSingleAnno(undefined)
     }
@@ -1065,10 +1040,6 @@ class Canvas extends Component{
                 this.onAnnoLabelInputClose()
             }
         }
-        // if(this.props.onAnnoSelect){
-        //     const anno = this.findAnno(annoId)
-        //     this.props.onAnnoSelect(anno)
-        // }
     }
 
     /**
@@ -1244,6 +1215,7 @@ class Canvas extends Component{
                     newAnno.id
                 )
             }
+            this.handleAnnoEvent(newAnno, canvasActions.ANNO_ENTER_CREATE_MODE)
         }
     }
 
@@ -1252,6 +1224,7 @@ class Canvas extends Component{
      * @param {string} id of annotation
      */
     recreateAnnotation(annoID) {
+        console.log('AnnoSave -> recreateAnnotation ', annoID)
 
         let annos = this.state.annos
 
@@ -1277,7 +1250,7 @@ class Canvas extends Component{
             type: anno.type,
             data: anno.data,
             mode: modes.CREATE,
-            status: (anno.status == 'database' ? annoStatus.CHANGED : annoStatus.NEW),
+            status: (anno.status === 'database' || anno.status === 'changed' ? annoStatus.CHANGED : annoStatus.NEW),
             labelIds: anno.labelIds,
             selectedNode: anno.data.length - 1,
             annoTime: anno.annoTime
@@ -1331,15 +1304,29 @@ class Canvas extends Component{
         return colorlut.getDefaultColor()
     }
 
+    updateAnnoBySaveResponse(res){
+        if (!res) return
+        if (res.tempId !== res.dbId){
+            const anno = this.findAnno(res.tempId)
+            anno.id = res.dbId
+            anno.status = annoStatus.DATABASE
+            this.updateSelectedAnno(anno)
+        } else {
+            const anno = this.findAnno(res.dbId)
+            anno.status = res.newStatus
+            this.updateSelectedAnno(anno)
+        }
+    }
+
     /**
      * Update selected anno and override mode if desired
      * 
-     * @param {object} anno - The new annotation the becomes the selected anno
+     * @param {object} anno - The new annotation that becomes the selected anno
      * @param {string} mode - The new mode for the selected anno
      * @returns The new anno that was set as selectedAnno in state and 
      *      the new annos list that was set in state
      */
-    updateSelectedAnno(anno, mode=undefined){
+    updateSelectedAnno(anno, mode=undefined, returnNewAnno=false){
         if (!anno) return
         const {newAnnos, newAnno} = this.mergeSelectedAnno(anno, mode)
         this.setState({
@@ -1347,12 +1334,11 @@ class Canvas extends Component{
             selectedAnnoId: anno.id,
             prevLabel: anno.labelIds, 
         })
-        // if(this.props.onAnnoSelect){
-        //     if (newAnno !== null){
-        //         this.props.onAnnoSelect(newAnno)
-        //     }
-        // }
-        return newAnnos
+        if (returnNewAnno){
+            return {newAnnos, newAnno}
+        } else {
+            return newAnnos
+        }
     }
 
     mergeSelectedAnno(anno, mode=undefined){
@@ -1396,13 +1382,6 @@ class Canvas extends Component{
     updateImageSize(){
         
         var container = this.props.container.current.getBoundingClientRect()
-        
-        // if (container.left < this.props.uiConfig.toolBarWidth){
-        //     canvasLeft = this.props.uiConfig.toolBarWidth + 10
-        // } else {
-        //     canvasLeft = container.left
-        // }
-        // var clientWidth = document.documentElement.clientWidth
         var clientHeight = document.documentElement.clientHeight
         var canvasTop
         var canvasLeft
@@ -1420,7 +1399,6 @@ class Canvas extends Component{
             maxImgHeight = clientHeight - container.top
             maxImgWidth = container.right -canvasLeft
         }
-        // if (this.props.appliedFullscreen) maxImgHeight = maxImgHeight + 10 
         var ratio = this.img.current.naturalWidth / this.img.current.naturalHeight
         var imgWidth = "100%"
         var imgHeight = "100%"
@@ -1471,9 +1449,6 @@ class Canvas extends Component{
 
     svgUpdate(svg){
         this.triggerCanvasEvent(canvasActions.CANVAS_SVG_UPDATE, svg)
-        // if(this.props.onSVGUpdate){
-        //     this.props.onSVGUpdate(svg)
-        // }
     }
 
     setImageLabels(labelIds){
@@ -1546,6 +1521,7 @@ class Canvas extends Component{
                     initLabelIds={this.state.imgLabelIds}
                     relatedId={this.props.imageMeta.id}
                     defaultLabel={this.props.defaultLabel}
+                    // onLabelConfirmed = {label => this.handleImgLabelUpdate(label)}
                     // disabled={!this.props.allowedActions.label}
                     // renderPopup
                 />
@@ -1585,17 +1561,8 @@ class Canvas extends Component{
             multilabels={this.props.canvasConfig.annos.multilabels}
             mousePos={this.mousePosAbs}
             defaultLabel={this.props.defaultLabel}
-                    // multilabels={true}
         />
     }
-
-    // renderAnnoCommentInput(selectedAnno){
-    //     return <AnnoCommentInput 
-    //         triggerInput={this.state.annoCommentInputTrigger}
-    //         onClose={comment => this.handleAnnoCommentInputClose(comment)}
-    //         iniComment={selectedAnno ? selectedAnno.comment : ''}
-    //     />
-    // }
 
     render(){
         const selectedAnno = this.findAnno(this.state.selectedAnnoId)
@@ -1678,7 +1645,6 @@ class Canvas extends Component{
                     onLoad={() => {this.onImageLoad()}} src={this.state.imageBlob}
                     width="100%" height="100%"
                 />
-                {/* </div> */}
                 </div>
                 {/* Placeholder for Layout*/}
                 <div style={{minHeight: this.state.svg.height}}></div>