+++ frontend/lost/src/containers/Annotation/SIA/lost-sia/CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased] -
+## [unreleased] - 
 ### Changed
 - Order of buttons in ToolBar: Move setting and filter btn up
+### Added 
+- onAnnoSaveEvent and annoSaveResponse for instant annotation update in backend
+### Removed 
+- autoSave Feature since it caused a lot of runtime errors and instability
 
 ## [1.3.0] - 2022-05-13
 ### Added